Limit garbage collection threshold to 800000
Based on:
http://bling.github.io/blog/2016/01/18/why-are-you-changing-gc-cons-threshold/
A reminder of the problem, in bling's own words:
The default value for gc-cons-threshold is quite low by modern standards – it’s only 800KB. But is that actually a problem? Emacs is known to have poor defaults, but most of the time defaults are chosen because they work most of the time, and something as crucial as gc-cons-threshold was probably chosen with at least some thought. But if that’s the case, then why is everyone (including some popular distributions) increasing the default value?